pyproject.toml
[project]
name = "reachability-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
]
